Species Variation Principal Component Analysis
PCA includes two functions, one of which creates a scree plot that plots the variance of principal components. The other creates a menu that the scree plot directly correlates with, allowing the user to pick which principal components they'd like to use as the x and y axis. The function plots all the metrics and gives the user the option to plot 95% percent confidence interval.